Moved Equinor.ProCoSys.Completion.Query.PunchItemQueries.GetPunchItemsByCheckList.GetPunchItemsByCheckListGuidQuery to Equinor.ProCoSys.Completion.Query.CheckListQueries.GetPunchItems.GetPunchItemsQuery
