+++ src/Equinor.ProCoSys.Completion.Query/CheckListQueries/GetPunchItems/GetPunchItemsQuery.cs
@@ -1,14 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
-using Equinor.ProCoSys.Completion.Domain;
 using Equinor.ProCoSys.Completion.Query.PunchItemServices;
 using MediatR;
 using ServiceResult;
 
-namespace Equinor.ProCoSys.Completion.Query.PunchItemQueries.GetPunchItemsByCheckList;
+namespace Equinor.ProCoSys.Completion.Query.CheckListQueries.GetPunchItems;
 
-public class GetPunchItemsByCheckListGuidQuery(Guid checkListGuid)
-    : INeedProjectAccess, IRequest<Result<IEnumerable<PunchItemDetailsDto>>>, IIsCheckListQuery
+public class GetPunchItemsQuery(Guid checkListGuid)
+    : IRequest<Result<IEnumerable<PunchItemDetailsDto>>>, IIsCheckListQuery
 {
     public Guid CheckListGuid { get; } = checkListGuid;
     public CheckListDetailsDto CheckListDetailsDto { get; set; } = null!;
