+++ src/Equinor.ProCoSys.Completion.Query/CheckListQueries/IIsCheckListQuery.cs
@@ -1,0 +1,11 @@
+﻿using Equinor.ProCoSys.Completion.Domain;
+using MediatR;
+using System;
+
+namespace Equinor.ProCoSys.Completion.Query.CheckListQueries;
+
+public interface IIsCheckListQuery : INeedProjectAccess, IBaseRequest
+{
+    Guid CheckListGuid { get; }
+    CheckListDetailsDto CheckListDetailsDto { get; set; }
+}
