--- src/Equinor.ProCoSys.Completion.Query/PunchItemQueries/IIsCheckListQuery.cs
@@ -1,10 +1,0 @@
-﻿using MediatR;
-using System;
-
-namespace Equinor.ProCoSys.Completion.Query.PunchItemQueries;
-
-public interface IIsCheckListQuery : IBaseRequest
-{
-    Guid CheckListGuid { get; }
-    CheckListDetailsDto CheckListDetailsDto { get; set; }
-}
